feat: allow setting components of `FriParameters` separately
Previously, the whole `FriParameters` had to be specified by the user to customize it. This change allows users to customize its components (fri_step_list, last_layer_degree_bound, etc) separately.

This helps isolate the behavior of `FriParameters.fri_step_list`, which doesn't have a default value and is autogenerated if not specified